reattempting fix hashtagrepo

+++ src/main/java/com/example/bitter/repository/HashtagRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HashtagRepository extends JpaRepository <Hashtag, Integer>{
+public interface HashtagRepository extends JpaRepository <Hashtag, Long>{
     Optional<Hashtag> findByLabel(String label);
 }
